dev-requirements.txt
mock
pytest-django
pytest
pytest-cov
tox
ruff
